requirements.txt
Django==4.2.5
django-filter==2.4.0
djangorestframework==3.14.0
python-dotenv==1.0.0
djoser==2.2.0